renames stats folder
