+++ src/stats/chart-view.js
