+++ src/stats/charts-model.js
